pyproject.toml
[tool.poetry]
name = "song-challenge"
version = "0.1.0"
description = ""
authors = ["Esteban Soler <esteban@informalthinkers.com>"]
license = "MIT"
readme = "README.md"
# packages = [{include = "song_challenge"}]

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.6.0"
certifi = "2022.12.7"
cffi = "1.15.1"
charset-normalizer = "3.0.1"
crispy-bootstrap5 = "0.7"
cryptography = "39.0.1"
defusedxml = "0.7.1"
django = "4.2"
django-allauth = "0.52.0"
django-crispy-forms = "2.0"
django-debug-toolbar = "3.8.1"
gunicorn = "20.1.0"
idna = "3.4"
oauthlib = "3.2.2"
psycopg = "3.1.8"
psycopg-binary = "3.1.8"
pycparser = "2.21"
pyjwt = "2.6.0"
python3-openid = "3.2.0"
requests = "2.28.2"
requests-oauthlib = "1.3.1"
sqlparse = "0.4.3"
typing-extensions = "4.5.0"
urllib3 = "1.26.14"
whitenoise = "6.4.0"
dj-database-url = "^1.3.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.3"
django-filter = "^23.1"
django-cors-headers = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
